+++ CHANGELOG.md
@@ -42,6 +42,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### API Breaking
+
+* [#709](https://github.com/NibiruChain/nibiru/pull/709) Add fields to `OpenPosition` response.
+* [#707](https://github.com/NibiruChain/nibiru/pull/707) Add fluctuation limit checks in vpool methods.
+* [#712](https://github.com/NibiruChain/nibiru/pull/712) Add funding rate calculation and `FundingRateChangedEvent`.
+
 ### Upgrades
 
 * [#702](https://github.com/NibiruChain/nibiru/pull/702) Add upgrade handler for v0.10.0.